AIDES-Windows/requirements.txt
psutil
pywin32
sqlite3-binary; platform_system == "Windows"
pyyaml
pytest
rich
scikit-learn
